Renamed all files with new game name
