+++ code/larcenygame/larcenygame.c
@@ -1,5 +1,5 @@
 /***************************************************************
-                         larcenygame.c
+                         rippergame.c
                                
 RiPpEr253's entry into the N64Brew 2024 Gamejam
 ***************************************************************/
@@ -7,29 +7,22 @@ RiPpEr253's entry into the N64Brew 2024 Gamejam
 #include <libdragon.h>
 #include "../../core.h"
 #include "../../minigame.h"
-#include "larcenygame.h"
-#include "larcenygameAI.h"
-// TODO: debug stuff
-#include <inttypes.h>
+#include "rippergame.h"
+#include "rippergameAI.h"
 
 #define COUNTDOWN_DELAY 4.0f
 #define FINISH_DELAY 10.0f
 #define STARTING_GAME_TIME 60.0f
 #define DEFAULT_ABILITY_COOLDOWN 2.0f
-#define GUARD_ABILITY_RANGE 50
-#define THIEF_ABILITY_RANGE 50
-#define THIEF_ABILITY_STRIDE 4.5f
-#define OBJECTIVE_TOUCH_DISTANCE 27
 
 #define CONTROLLER_LOWER_DEADZONE 4
 #define CONTROLLER_UPPER_DEADZONE 50
-#define CONTROLLER_DEFAULT_DEBOUNCE 0.5f
-
-#define HUD_SAFE_AREA_X 22.0f
-#define HUD_SAFE_AREA_Y 28.0f
 
 #define FONT_DEBUG 1
 #define FONT_BILLBOARD 2
+
+#define RESOLUTION_WIDTH 320
+#define RESOLUTION_HEIGHT 240
 
 #define DEFAULT_PLAYER_SCALE 0.375f
 
@@ -55,16 +48,9 @@ effect_data effectPool[MAXPLAYERS];
 objective_data objectives[2];
 collisionobject_data collisionObjects[9];
 
-// camera variables
-cameraanimation_data cameraIntroKeyframes[10];
-cameraanimation_data cameraOutroKeyframes[2];
-animatedcameraobject_data animatedCamera;
-
 // drawing variables
 rspq_block_t* dplMap;
 rspq_syncpoint_t syncPoint;
-int resolutionDisplayX;
-int resolutionDisplayY;
 
 // fonts
 rdpq_font_t* fontDebug;
@@ -76,7 +62,6 @@ wav64_t sfx_countdown;
 wav64_t sfx_winner;
 wav64_t sfx_objectiveCompleted;
 wav64_t sfx_guardStunAbility;
-wav64_t sfx_guardStunAbilityHit;
 wav64_t sfx_thiefJumpAbility;
 wav64_t sfx_thiefCaught;
 
@@ -87,8 +72,6 @@ int lastCountdownNumber;
 float countdownTimer;
 bool gameStarting;
 bool gameEnding;
-bool gamePaused;
-float gamePauseDebounce;
 float gameTimeRemaining;
 
 player_team winningTeam;
@@ -96,22 +79,12 @@ player_team winningTeam;
 // You need this function defined somewhere in your project
 // so that the minigame manager can work
 const MinigameDef minigame_def = {
-    .gamename = "Larceny",
-    .developername = "RiPpEr253, music by drozerix",
-    .description = "A 2v2 game pitting Thieves against Guards and their treasures, hold r on start for hires",
-    .instructions = "Stick: Move, A: Thief: Wall Jump, Guard: Stun Thief"
+    .gamename = "RiPpEr's Game",
+    .developername = "RiPpEr253",
+    .description = "This is an example game.",
+    .instructions = "Mash A to win."
 };
 
-// Debug accumulators
-uint64_t startTime, endTime;
-uint64_t aiTime = 0;
-uint64_t fixedUpdateTime = 0;
-uint64_t updateTime = 0;
-uint64_t colTime = 0;
-uint64_t playerUpdate = 0;
-uint64_t drawTime = 0;
-uint64_t subDrawTime = 0;
-uint64_t totalFrameTime = 0;
 
 //TODO remove:
 float tempSpeed;
@@ -127,221 +100,43 @@ void debugInfoDraw(float deltaTime)
     rdpq_sync_pipe(); // Hardware crashes otherwise
     rdpq_sync_tile(); // Hardware crashes otherwise
 
-    rdpq_textparms_t textparms = { .align = ALIGN_LEFT, .width = resolutionDisplayX, .disable_aa_fix = true, };
+    rdpq_textparms_t textparms = { .align = ALIGN_LEFT, .width = RESOLUTION_WIDTH, .disable_aa_fix = true, };
     //rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+10, "Test Debug");
-    rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
     rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+20, "FPS: %f", 1.0f/deltaTime);
-
-    rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+30, "\
-    startTime=%"PRId64"\n\
-    endTime=%"PRId64"\n\
-    aiTime=%"PRId64"\n\
-    updateTime=%"PRId64"\n\
-    colTime=%"PRId64"\n\
-    fixedUpdate=%"PRId64"\n\
-    playerUpdate=%"PRId64"\n\
-    drawTime=%"PRId64"\n\
-    subDrawTime=%"PRId64"\n\
-    totalFrameTime=%"PRId64"\n",
-    startTime,
-    endTime,
-    aiTime,
-    updateTime,
-    colTime,
-    fixedUpdateTime,
-    playerUpdate,
-    drawTime,
-    subDrawTime,
-    totalFrameTime);
     //rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+30, "AI Statuses: P1: %i P2: %i P3: %i P4: %i", ai_getCurrentStateAsInt(0), ai_getCurrentStateAsInt(1), ai_getCurrentStateAsInt(2), ai_getCurrentStateAsInt(3));
     //rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+40, "Intersection point: %f, %f", tempIntersectionPoint.x, tempIntersectionPoint.z);
     //rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+50, "p1 x: %f, p1 y: %f", players[0].playerPos.v[0], players[0].playerPos.v[2]);
-    //rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+60, "Debounce Status: %f", gamePauseDebounce);
-}
-
-/*==============================
-    cameraAnimation_init
-    ensures there's sane defaults in the animated camera
-==============================*/
-
-void cameraAnimation_init()
-{
-    animatedCamera.currentAnimationLength = 0;
-    animatedCamera.currentAnimationKeyframe = 0;
-    animatedCamera.currentAnimationTime = 0.0f;
-    animatedCamera.currentlyPlaying = false;
-}
-
-/*==============================
-    startCameraAnimation
-    takes in a reference to a camera data array and 
-    sets up the animated camera to play the animation
-==============================*/
-
-void startCameraAnimation(cameraanimation_data* newAnimation, int lengthOfAnimation)
-{
-    animatedCamera.currentAnimation = newAnimation;
-    animatedCamera.currentAnimationLength = lengthOfAnimation;
-    animatedCamera.currentAnimationKeyframe = 0;
-    animatedCamera.currentAnimationTime = 0.0f;
-    animatedCamera.currentlyPlaying = true;
-}
-
-/*==============================
-    cameraAnimation_update
-    updates the animated camera if there's any more keyframes left in it
-==============================*/
-
-void cameraAnimation_update(float deltaTime)
-{
-    // if not currently playing, return as pointers will be invalid
-    if(!animatedCamera.currentlyPlaying)
+    /*joypad_port_t controllerPort = core_get_playercontroller(0);
+    
+    if(!joypad_is_connected(controllerPort))
     {
         return;
     }
-
-    // advance the time
-    animatedCamera.currentAnimationTime += deltaTime;
-
-    // if time elapsed > keyframe length, then it's time for the next frame
-    if(animatedCamera.currentAnimationTime >= animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].timeUntilNextKeyframe)
+    else
     {
-        animatedCamera.currentAnimationTime = 0.0f;
-        animatedCamera.currentAnimationKeyframe += 1;
-    }
+        joypad_inputs_t joypad = joypad_get_inputs(controllerPort);
+        rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+40, "p1 stick x: %i, p1 stick y: %i, p1 speed: %f", joypad.stick_x, joypad.stick_y, tempSpeed);
+        
+        // TODO: remove
+        // linear algebra tests
+        T3DVec3 tempvec = {0};
+        tempvec = (T3DVec3){{0,0,1}};
+        tempvec.v[0] = -sinf(players[0].rotY);
+        tempvec.v[2] = cosf(players[0].rotY);
 
-    // if we've reached the end of the animation, set not playing at return
-    if(animatedCamera.currentAnimationKeyframe >= animatedCamera.currentAnimationLength)
-    {
-        // if an animation is over, make sure the exact end points are set to avoid high deltatimes making the camera crooked
-        camPos = animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe - 1].camEndPos;
-        camTarget = animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe - 1].lookAtEnd;
-        animatedCamera.currentlyPlaying = false;
-        return;
-    }
+        t3d_vec3_add(&tempvec, &tempvec, &players[0].playerPos);
 
-    // use linear interpolation for all three positions for the camera
-    t3d_vec3_lerp( &camPos,     &animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].camStartPos, 
-                                &animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].camEndPos, 
-                                animatedCamera.currentAnimationTime / animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].timeUntilNextKeyframe);
-    // do the same for the target
-    t3d_vec3_lerp(&camTarget,   &animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].lookAtStart, 
-                                &animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].lookAtEnd, 
-                                animatedCamera.currentAnimationTime / animatedCamera.currentAnimation[animatedCamera.currentAnimationKeyframe].timeUntilNextKeyframe);
-
-}
-
-/*==============================
-    startCameraAnimationIntro
-    populates an animation for the intro and then sends it to the camera animator object
-==============================*/
-
-void startCameraAnimationIntro()
-{
-    // start high and move to centre
-    cameraIntroKeyframes[0].camStartPos = (T3DVec3){{0.0f, 511.0f, 45.0f}};
-    cameraIntroKeyframes[0].camEndPos = (T3DVec3){{0.0f, 255.0f, 45.0f}};
-    cameraIntroKeyframes[0].lookAtStart = (T3DVec3){{0, 0, -5}};
-    cameraIntroKeyframes[0].lookAtEnd = (T3DVec3){{0, 0, -5}};
-    cameraIntroKeyframes[0].timeUntilNextKeyframe = 0.5f;
-
-    // look at objective 1
-    cameraIntroKeyframes[1].camStartPos = (T3DVec3){{0.0f, 255.0f, 45.0f}};
-    cameraIntroKeyframes[1].camEndPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[1].lookAtStart = (T3DVec3){{0, 0, -5}};
-    cameraIntroKeyframes[1].lookAtEnd = objectives[0].objectivePos;
-    cameraIntroKeyframes[1].timeUntilNextKeyframe = 0.5f;
-    
-    // hold camera
-    cameraIntroKeyframes[2].camStartPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[2].camEndPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[2].lookAtStart = objectives[0].objectivePos;
-    cameraIntroKeyframes[2].lookAtEnd = objectives[0].objectivePos;
-    cameraIntroKeyframes[2].timeUntilNextKeyframe = 1.0f;
-    
-    // look at objective 2
-    cameraIntroKeyframes[3].camStartPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[3].camEndPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[3].lookAtStart = objectives[0].objectivePos;
-    cameraIntroKeyframes[3].lookAtEnd = objectives[1].objectivePos;
-    cameraIntroKeyframes[3].timeUntilNextKeyframe = 0.5f;
-    
-    // hold camera
-    cameraIntroKeyframes[4].camStartPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[4].camEndPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[4].lookAtStart = objectives[1].objectivePos;
-    cameraIntroKeyframes[4].lookAtEnd = objectives[1].objectivePos;
-    cameraIntroKeyframes[4].timeUntilNextKeyframe = 1.0f;
-    
-    // look at player 1
-    cameraIntroKeyframes[5].camStartPos = (T3DVec3){{0.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[5].camEndPos = (T3DVec3){{-105.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[5].lookAtStart = objectives[1].objectivePos;
-    cameraIntroKeyframes[5].lookAtEnd = players[0].playerPos;
-    cameraIntroKeyframes[5].timeUntilNextKeyframe = 0.5f;
-
-    // hold camera
-    cameraIntroKeyframes[6].camStartPos = (T3DVec3){{-105.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[6].camEndPos = (T3DVec3){{-105.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[6].lookAtStart = players[0].playerPos;
-    cameraIntroKeyframes[6].lookAtEnd = players[0].playerPos;
-    cameraIntroKeyframes[6].timeUntilNextKeyframe = 1.0f;
-    
-    // look at player 4
-    cameraIntroKeyframes[7].camStartPos = (T3DVec3){{-105.0f, 90.0f, 45.0f}};
-    cameraIntroKeyframes[7].camEndPos = (T3DVec3){{-105.0f, 90.0f, -45.0f}};
-    cameraIntroKeyframes[7].lookAtStart = players[0].playerPos;
-    cameraIntroKeyframes[7].lookAtEnd = players[3].playerPos;
-    cameraIntroKeyframes[7].timeUntilNextKeyframe = 0.5f;
-
-    // hold camera
-    cameraIntroKeyframes[8].camStartPos = (T3DVec3){{-105.0f, 90.0f, -45.0f}};
-    cameraIntroKeyframes[8].camEndPos = (T3DVec3){{-105.0f, 90.0f, -45.0f}};
-    cameraIntroKeyframes[8].lookAtStart = players[3].playerPos;
-    cameraIntroKeyframes[8].lookAtEnd = players[3].playerPos;
-    cameraIntroKeyframes[8].timeUntilNextKeyframe = 1.0f;
-
-    // return to centre
-    cameraIntroKeyframes[9].camStartPos = (T3DVec3){{-105.0f, 90.0f, -45.0f}};
-    cameraIntroKeyframes[9].camEndPos = (T3DVec3){{0.0f, 255.0f, 45.0f}};
-    cameraIntroKeyframes[9].lookAtStart = players[3].playerPos;
-    cameraIntroKeyframes[9].lookAtEnd = (T3DVec3){{0, 0, -5}};
-    cameraIntroKeyframes[9].timeUntilNextKeyframe = 0.5f;
-
-    startCameraAnimation(cameraIntroKeyframes, 10);
-}
-
-/*==============================
-    startCameraAnimationOutro
-    populates an animation for the outro and then sends it to the camera animator object
-==============================*/
-
-void startCameraAnimationOutro()
-{
-    // hold camera
-    cameraOutroKeyframes[0].camStartPos = (T3DVec3){{0, 255.0f, 45.0f}};
-    cameraOutroKeyframes[0].camEndPos = (T3DVec3){{0, 255.0f, 45.0f}};
-    cameraOutroKeyframes[0].lookAtStart = (T3DVec3){{0, 0, -5}};
-    cameraOutroKeyframes[0].lookAtEnd = (T3DVec3){{0, 0, -5}};
-    cameraOutroKeyframes[0].timeUntilNextKeyframe = 1.5f;
-    // zoom through map
-    cameraOutroKeyframes[1].camStartPos = (T3DVec3){{0, 255.0f, 45.0f}};
-    cameraOutroKeyframes[1].camEndPos = (T3DVec3){{0, 0.0f, 0.0f}};
-    cameraOutroKeyframes[1].lookAtStart = (T3DVec3){{0, 0, -5}};
-    cameraOutroKeyframes[1].lookAtEnd = (T3DVec3){{0, -150, -5}};
-    cameraOutroKeyframes[1].timeUntilNextKeyframe = 0.5f;
-
-    startCameraAnimation(cameraOutroKeyframes, 2);
+        rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 10, 80+50, "rotation of player: %f, vector position test X: %f, Z: %f",players[0].rotY,tempvec.x, tempvec.z);
+    }*/
 }
 
 /*==============================
     HUD_Update
     updates HUD elements outside of the draw loop
 ==============================*/
-
 void HUD_Update(float deltaTime)
 {
-    if(!gameStarting && !gameEnding && !gamePaused)
+    if(!gameStarting && !gameEnding)
     {
         // Update anything needed on the HUD (ability timer bar resizing/game time remaining)
         gameTimeRemaining -= deltaTime;
@@ -352,14 +147,13 @@ void HUD_Update(float deltaTime)
     HUD_draw
     draws main HUD info using the RDP
 ==============================*/
-
 void HUD_draw()
 {
     const T3DVec3 HUDOffsets[] = {
-    (T3DVec3){{HUD_SAFE_AREA_X, resolutionDisplayY - HUD_SAFE_AREA_Y, 0.0f}},
-    (T3DVec3){{resolutionDisplayX - HUD_SAFE_AREA_X - 80.0f, resolutionDisplayY - HUD_SAFE_AREA_Y, 0.0f}},
-    (T3DVec3){{resolutionDisplayX - HUD_SAFE_AREA_X - 80.0f, HUD_SAFE_AREA_Y, 0.0f}},
-    (T3DVec3){{HUD_SAFE_AREA_X, HUD_SAFE_AREA_Y, 0.0f}},
+    (T3DVec3){{16.0f, 208.0f, 0.0f}},
+    (T3DVec3){{RESOLUTION_WIDTH - 96.0f, 208.0f, 0.0f}},
+    (T3DVec3){{RESOLUTION_WIDTH - 96.0f, 16.0f, 0.0f}},
+    (T3DVec3){{16.0f, 16.0f, 0.0f}},
     };
 
     const color_t colours[] = {
@@ -369,16 +163,12 @@ void HUD_draw()
     PLAYERCOLOR_4,
     };
 
-    rdpq_textparms_t textparms = {.style_id = 7, .align = ALIGN_CENTER, .width = resolutionDisplayX, .disable_aa_fix = true, };
-    rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
+    // make sure the RDP is sync'd
+    rdpq_sync_tile();
+    rdpq_sync_pipe(); // Hardware crashes otherwise
+
+    rdpq_textparms_t textparms = { .align = ALIGN_CENTER, .width = RESOLUTION_WIDTH, .disable_aa_fix = true, };
     rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, 20, "Time Left: %i", (int)gameTimeRemaining);
-
-    // TODO: Debug stuff
-    //uint64_t subDrawStart = get_ticks();
-
-    
-    rdpq_textparms_t playerHUDTextParms = {.align = ALIGN_LEFT, .indent = 20, .line_spacing = -1, .style_id = 0, .disable_aa_fix = true, };
-
 
     // iterate through all the players and draw what we need
     for(int iDx = 0; iDx < MAXPLAYERS; iDx++)
@@ -388,13 +178,11 @@ void HUD_draw()
         {
             continue;
         }
-
-        playerHUDTextParms.style_id = iDx;
-
+        
         // set up some basic vectors for reverse picking
         T3DVec3 billboardPos = (T3DVec3){{
             players[iDx].playerPos.v[0],
-            players[iDx].playerPos.v[1]+10.0f,
+            players[iDx].playerPos.v[1],
             players[iDx].playerPos.v[2]
         }};
 
@@ -409,35 +197,33 @@ void HUD_draw()
         rdpq_sync_tile();
         rdpq_sync_pipe(); // Hardware crashes otherwise
 
-        //rdpq_text_printf(&playerHUDTextParms, FONT_BILLBOARD, x-5, y-16, "P%d", iDx+1);
+        rdpq_text_printf(&(rdpq_textparms_t){ .style_id = iDx, .disable_aa_fix = true, }, FONT_BILLBOARD, x-5, y-16, "P%d", iDx+1);
         
         if(players[iDx].stunTimer > 0.0f)
         {
-            rdpq_text_printf(&playerHUDTextParms, FONT_BILLBOARD, x-25, y-16, "Stunned: %.2f", players[iDx].stunTimer);
-        }
-        else
-        {
-            rdpq_text_printf(&playerHUDTextParms, FONT_BILLBOARD, x-25, y-16, "P%d", iDx+1);
+            rdpq_text_printf(&(rdpq_textparms_t){ .style_id = iDx, .disable_aa_fix = true, }, FONT_BILLBOARD, x-35, y-26, "Stunned: %.2f", players[iDx].stunTimer);
         }
 
         // draw the rest of the text for the HUD
+        // draw the A button sprite
         rdpq_set_mode_standard();
         rdpq_mode_alphacompare(128);
         rdpq_sync_load();
         
-        rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
+        // make sure the RDP is sync'd
+        rdpq_sync_tile();
+        rdpq_sync_pipe(); // Hardware crashes otherwise
 
-        // draw the A button sprite
-        rdpq_sprite_blit(spriteAButton, HUDOffsets[iDx].v[0], HUDOffsets[iDx].v[1] - 16, &(rdpq_blitparms_t){ .scale_x = 1.0f,  .scale_y = 1.0f});
+        rdpq_sprite_blit(spriteAButton, HUDOffsets[iDx].v[0], HUDOffsets[iDx].v[1] - 8, &(rdpq_blitparms_t){ .scale_x = 0.5f,  .scale_y = 0.5f});
         if(players[iDx].playerTeam == teamThief)
         {
-            rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
-            rdpq_text_printf(&playerHUDTextParms, FONT_BILLBOARD, HUDOffsets[iDx].v[0], HUDOffsets[iDx].v[1]-5, "Jump Wall\nPlayer %i: Thief", iDx + 1);
+            rdpq_text_printf(&(rdpq_textparms_t){ .style_id = iDx, .disable_aa_fix = true, }, FONT_BILLBOARD, HUDOffsets[iDx].v[0] + 24, HUDOffsets[iDx].v[1], "Jump Wall");
+            rdpq_text_printf(&(rdpq_textparms_t){ .style_id = iDx, .disable_aa_fix = true, }, FONT_BILLBOARD, HUDOffsets[iDx].v[0], HUDOffsets[iDx].v[1] + 10, "Player %i: Thief", iDx + 1);
         }
         else if(players[iDx].playerTeam == teamGuard)
-        { 
-            rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
-            rdpq_text_printf(&playerHUDTextParms, FONT_BILLBOARD, HUDOffsets[iDx].v[0], HUDOffsets[iDx].v[1]-5, "Stun Attack\nPlayer %i: Guard", iDx + 1);
+        {
+            rdpq_text_printf(&(rdpq_textparms_t){ .style_id = iDx, .disable_aa_fix = true, }, FONT_BILLBOARD, HUDOffsets[iDx].v[0] + 24, HUDOffsets[iDx].v[1], "Stun Attack");
+            rdpq_text_printf(&(rdpq_textparms_t){ .style_id = iDx, .disable_aa_fix = true, }, FONT_BILLBOARD, HUDOffsets[iDx].v[0], HUDOffsets[iDx].v[1] + 10, "Player %i: Guard", iDx + 1);
         }
 
         // make sure the RDP is sync'd
@@ -452,9 +238,6 @@ void HUD_draw()
         rdpq_fill_rectangle(HUDOffsets[iDx].v[0] - 2, HUDOffsets[iDx].v[1] + 12, HUDOffsets[iDx].v[0] + t3d_lerp(85, -2, (players[iDx].abilityTimer / DEFAULT_ABILITY_COOLDOWN) ) , HUDOffsets[iDx].v[1] + 15);
         
     }
-
-    // TODO: Debug stuff
-    //subDrawTime += get_ticks() - subDrawStart;
 }
 
 /*==============================
@@ -462,14 +245,13 @@ void HUD_draw()
     Initialises collision objects array manually
     with positions, sizes and collision types
 ==============================*/
-
 void collision_init()
 {
     collisionObjects[0].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
     collisionObjects[0].collisionCentrePos = (T3DVec3){{0.0f, 1.0f, 0.0f}};
     collisionObjects[0].collisionType = collisionAll;
-    collisionObjects[0].sizeX = 120;
-    collisionObjects[0].sizeZ = 40;
+    collisionObjects[0].sizeX = 100;
+    collisionObjects[0].sizeZ = 20;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[0].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -477,10 +259,10 @@ void collision_init()
     collisionObjects[0].dplCollision = rspq_block_end();
     
     collisionObjects[1].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[1].collisionCentrePos = (T3DVec3){{-75.0f, 1.0f, 106.0f}};
+    collisionObjects[1].collisionCentrePos = (T3DVec3){{-80.0f, 1.0f, 106.0f}};
     collisionObjects[1].collisionType = collisionAll;
-    collisionObjects[1].sizeX = 80;
-    collisionObjects[1].sizeZ = 22;
+    collisionObjects[1].sizeX = 60;
+    collisionObjects[1].sizeZ = 12;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[1].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -488,10 +270,10 @@ void collision_init()
     collisionObjects[1].dplCollision = rspq_block_end();
 
     collisionObjects[2].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[2].collisionCentrePos = (T3DVec3){{-106.0f, 1.0f, -75.0f}};
+    collisionObjects[2].collisionCentrePos = (T3DVec3){{-106.0f, 1.0f, -80.0f}};
     collisionObjects[2].collisionType = collisionAll;
-    collisionObjects[2].sizeX = 22;
-    collisionObjects[2].sizeZ = 75;
+    collisionObjects[2].sizeX = 12;
+    collisionObjects[2].sizeZ = 60;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[2].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -499,10 +281,10 @@ void collision_init()
     collisionObjects[2].dplCollision = rspq_block_end();
 
     collisionObjects[3].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[3].collisionCentrePos = (T3DVec3){{-75.0f, 1.0f, -106.0f}};
+    collisionObjects[3].collisionCentrePos = (T3DVec3){{-80.0f, 1.0f, -106.0f}};
     collisionObjects[3].collisionType = collisionAll;
-    collisionObjects[3].sizeX = 80;
-    collisionObjects[3].sizeZ = 22;
+    collisionObjects[3].sizeX = 60;
+    collisionObjects[3].sizeZ = 12;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[3].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -510,10 +292,10 @@ void collision_init()
     collisionObjects[3].dplCollision = rspq_block_end();
 
     collisionObjects[4].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[4].collisionCentrePos = (T3DVec3){{106.0f, 1.0f, -75.0f}};
+    collisionObjects[4].collisionCentrePos = (T3DVec3){{106.0f, 1.0f, -80.0f}};
     collisionObjects[4].collisionType = collisionAll;
-    collisionObjects[4].sizeX = 22;
-    collisionObjects[4].sizeZ = 75;
+    collisionObjects[4].sizeX = 12;
+    collisionObjects[4].sizeZ = 60;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[4].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -521,10 +303,10 @@ void collision_init()
     collisionObjects[4].dplCollision = rspq_block_end();
 
     collisionObjects[5].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[5].collisionCentrePos = (T3DVec3){{75.0f, 1.0f, -106.0f}};
+    collisionObjects[5].collisionCentrePos = (T3DVec3){{80.0f, 1.0f, -106.0f}};
     collisionObjects[5].collisionType = collisionAll;
-    collisionObjects[5].sizeX = 80;
-    collisionObjects[5].sizeZ = 22;
+    collisionObjects[5].sizeX = 60;
+    collisionObjects[5].sizeZ = 12;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[5].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -532,10 +314,10 @@ void collision_init()
     collisionObjects[5].dplCollision = rspq_block_end();
 
     collisionObjects[6].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[6].collisionCentrePos = (T3DVec3){{106.0f, 1.0f, 75.0f}};
+    collisionObjects[6].collisionCentrePos = (T3DVec3){{106.0f, 1.0f, 80.0f}};
     collisionObjects[6].collisionType = collisionAll;
-    collisionObjects[6].sizeX = 22;
-    collisionObjects[6].sizeZ = 75;
+    collisionObjects[6].sizeX = 12;
+    collisionObjects[6].sizeZ = 60;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[6].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -543,10 +325,10 @@ void collision_init()
     collisionObjects[6].dplCollision = rspq_block_end();
 
     collisionObjects[7].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[7].collisionCentrePos = (T3DVec3){{75.0f, 1.0f, 106.0f}};
+    collisionObjects[7].collisionCentrePos = (T3DVec3){{80.0f, 1.0f, 106.0f}};
     collisionObjects[7].collisionType = collisionAll;
-    collisionObjects[7].sizeX = 80;
-    collisionObjects[7].sizeZ = 22;
+    collisionObjects[7].sizeX = 60;
+    collisionObjects[7].sizeZ = 12;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[7].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -554,10 +336,10 @@ void collision_init()
     collisionObjects[7].dplCollision = rspq_block_end();
 
     collisionObjects[8].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    collisionObjects[8].collisionCentrePos = (T3DVec3){{-106.0f, 1.0f, 75.0f}};
+    collisionObjects[8].collisionCentrePos = (T3DVec3){{-106.0f, 1.0f, 80.0f}};
     collisionObjects[8].collisionType = collisionAll;
-    collisionObjects[8].sizeX = 22;
-    collisionObjects[8].sizeZ = 75;
+    collisionObjects[8].sizeX = 12;
+    collisionObjects[8].sizeZ = 60;
         rspq_block_begin();
         t3d_matrix_push(collisionObjects[8].modelMatFP);
         t3d_model_draw(modelCollision);
@@ -572,7 +354,6 @@ void collision_init()
     Could be more optimised with display lists,
     but it's intended for debugging
 ==============================*/
-
 void collision_draw()
 {
     int numberOfObjects = sizeof(collisionObjects) / sizeof(collisionObjects[0]);
@@ -592,17 +373,17 @@ void collision_draw()
     }
 }
 
+// pass in position to check if intersecting
+// adjust the passed in vector accordingly
+// return true if colliding
+
 /*==============================
     collision_check
     pass in a collisionresult_data struct
     and a position to check if intersecting
 ==============================*/
-
 void collision_check(collisionresult_data* returnStruct, T3DVec3* pos)
 {
-    // TODO: debug stuff
-    //uint64_t colStart = get_ticks();
-
     returnStruct->didCollide = false; returnStruct->collisionType = collisionAll; returnStruct->indexOfCollidedObject = 0; returnStruct->intersectionPoint = (T3DVec3){{0}};
 
     int numberOfObjects = sizeof(collisionObjects) / sizeof(collisionObjects[0]);
@@ -616,14 +397,9 @@ void collision_check(collisionresult_data* returnStruct, T3DVec3* pos)
             pos->v[2] < collisionObjects[iDx].collisionCentrePos.v[2] + (collisionObjects[iDx].sizeZ / 2))
         {
             returnStruct->didCollide = true; returnStruct->collisionType = collisionObjects[iDx].collisionType; returnStruct->indexOfCollidedObject = iDx;
-            // TODO: debug stuff
-            //colTime += get_ticks() - colStart;
             return;
         }
     }
-
-    // TODO: debug stuff
-    //colTime += get_ticks() - colStart;
     return;
 }
 
@@ -656,11 +432,8 @@ bool lineLineIntersectTest(T3DVec3* AB, T3DVec3* CD, T3DVec3* XZ)
         return true;
     }
 }
-/*==============================
-    collision_check_intersect
-    a version of collision check that handles intersecting points
-==============================*/
 
+// a version of collision check that handles intersecting points
 void collision_check_intersect(collisionresult_data* returnStruct, T3DVec3* startingPos, T3DVec3* endingPos)
 {
     returnStruct->didCollide = false; returnStruct->collisionType = collisionAll; returnStruct->indexOfCollidedObject = 0; returnStruct->intersectionPoint = (T3DVec3){{0}};
@@ -675,60 +448,48 @@ void collision_check_intersect(collisionresult_data* returnStruct, T3DVec3* star
             endingPos->v[2] > collisionObjects[iDx].collisionCentrePos.v[2] - (collisionObjects[iDx].sizeZ / 2) &&
             endingPos->v[2] < collisionObjects[iDx].collisionCentrePos.v[2] + (collisionObjects[iDx].sizeZ / 2))
         {
-            // did collide, now figure out which side
-            // need an escape check by checking for collision again on the modified point?
-            // if starting pos X is lesser outside and ending pos X is inside, then hit on left side
-            if( startingPos->v[0] < collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeX / 2) &&
-                endingPos->v[0] > collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeX / 2))
-            {
-                // hit the left side
-                // set the X to the box wall, and the Z to the normal end point
-                returnStruct->intersectionPoint.v[0] = startingPos->v[0];
-                returnStruct->intersectionPoint.v[2] = endingPos->v[2];
-            }
-            // if starting pos X is greater and outside and ending pos X is inside, then hit on right side
-            if( startingPos->v[0] > collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeX / 2) &&
-                endingPos->v[0] < collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeX / 2))
-            {
-                // hit the right side
-                // set the X to the box wall, and the Z to the normal end point
-                returnStruct->intersectionPoint.v[0] = startingPos->v[0];
-                returnStruct->intersectionPoint.v[2] = endingPos->v[2];
-            }
-            // if starting pos Z is lesser and outside and ending pos X is inside, then hit on right side
-            if( startingPos->v[2] < collisionObjects[iDx].collisionCentrePos.v[2] - (collisionObjects[iDx].sizeZ / 2) &&
-                endingPos->v[2] > collisionObjects[iDx].collisionCentrePos.v[2] - (collisionObjects[iDx].sizeZ / 2))
-            {
-                // hit the top side
-                // set the Z to the box wall, and the X to the normal end point
-                returnStruct->intersectionPoint.v[0] = endingPos->v[0];
-                returnStruct->intersectionPoint.v[2] = startingPos->v[2];
-            }
-            // if starting pos Z is greater and outside and ending pos X is inside, then hit on right side
-            if( startingPos->v[2] > collisionObjects[iDx].collisionCentrePos.v[2] + (collisionObjects[iDx].sizeZ / 2) &&
-                endingPos->v[2] < collisionObjects[iDx].collisionCentrePos.v[2] + (collisionObjects[iDx].sizeZ / 2))
-            {
-                // hit the bottom side
-                // set the Z to the box wall, and the X to the normal end point
-                returnStruct->intersectionPoint.v[0] = endingPos->v[0];
-                returnStruct->intersectionPoint.v[2] = startingPos->v[2];
-            }
-
-            // make sure the intersection point, if any, makes sense
-            if(!(returnStruct->intersectionPoint.v[0] == 0.0f && returnStruct->intersectionPoint.v[2] == 0.0f))
-            {
-                collisionresult_data tempCollisionInfo;
-                collision_check(&tempCollisionInfo, &returnStruct->intersectionPoint);
-                if(tempCollisionInfo.didCollide)
-                {
-                    returnStruct->intersectionPoint = *startingPos;
-                }
-            }
-
+            // fill return struct as normal
             returnStruct->didCollide = true; returnStruct->collisionType = collisionObjects[iDx].collisionType; returnStruct->indexOfCollidedObject = iDx;
+            // determine the intersection point of the box
+            T3DVec3 IntersectionLine[2]; // line made up of collision checker movement
+            IntersectionLine[0] = *endingPos; IntersectionLine[1] = *startingPos;
+            // create lines of each of the four sides
+            T3DVec3 Line1[2]; // top line
+            // use mini arrays to determine the 4 points, counter clockwise wrapped starting top-right
+            Line1[0].x = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeX / 2); Line1[0].z = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeZ / 2);
+            Line1[1].x = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeX / 2); Line1[1].z = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeZ / 2);
+            T3DVec3 Line2[2]; // left line
+            Line2[0].x = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeX / 2); Line2[0].z = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeZ / 2);
+            Line2[1].x = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeX / 2); Line2[1].z = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeZ / 2);
+            T3DVec3 Line3[2]; // bottom line
+            Line3[0].x = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeX / 2); Line3[0].z = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeZ / 2);
+            Line3[1].x = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeX / 2); Line3[1].z = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeZ / 2);
+            T3DVec3 Line4[2]; // right line
+            Line4[0].x = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeX / 2); Line4[0].z = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeZ / 2);
+            Line4[1].x = collisionObjects[iDx].collisionCentrePos.v[0] + (collisionObjects[iDx].sizeX / 2); Line4[1].z = collisionObjects[iDx].collisionCentrePos.v[0] - (collisionObjects[iDx].sizeZ / 2);
+
+            T3DVec3 intersectionResult;
+            if(lineLineIntersectTest(Line1, IntersectionLine, &intersectionResult))
+            {
+                returnStruct->intersectionPoint = intersectionResult;
+            }
+            else if(lineLineIntersectTest(Line2, IntersectionLine, &intersectionResult))
+            {
+                returnStruct->intersectionPoint = intersectionResult;
+            }
+            else if(lineLineIntersectTest(Line3, IntersectionLine, &intersectionResult))
+            {
+                returnStruct->intersectionPoint = intersectionResult;
+            }
+            else if(lineLineIntersectTest(Line4, IntersectionLine, &intersectionResult))
+            {
+                returnStruct->intersectionPoint = intersectionResult;
+            }
+
             return;
         }
     }
+    return;
 }
 
 /*==============================
@@ -743,7 +504,6 @@ void collision_cleanup()
     for(int iDx = 0; iDx < numberOfObjects; iDx++)
     {
         free_uncached(collisionObjects[iDx].modelMatFP);
-        rspq_block_free(collisionObjects[iDx].dplCollision);
     }
 }
 
@@ -751,6 +511,7 @@ void collision_cleanup()
 /*==============================
     end_game
     Sets the game state to ending and performs any tasks that are related
+    
     Call when game is entirely over as this starts a timer to send us back to the main menu
 
     @param the winning team passed in as an enum
@@ -766,16 +527,8 @@ void end_game(player_team victoriousTeam)
     // set game state to game ending
     gameEnding = true;
     gameStarting = false;
-    gamePaused = false;
-    gamePauseDebounce = 0.0f;
-
-    xm64player_stop(&xm_music);
 
     wav64_play(&sfx_winner, 31);
-    mixer_ch_set_vol(31, 0.75f, 0.75f);
-
-    // set off the outro camera animation
-    startCameraAnimationOutro();
 
     // interate through teams and set a winner
     for(int iDx = 0; iDx < MAXPLAYERS; iDx++)
@@ -813,31 +566,27 @@ void player_init(int playerNumber)
     if(playerNumber == 0 || playerNumber == 2)
     {
         players[playerNumber].playerTeam = teamThief;
-        players[playerNumber].model = t3d_model_load("rom:/larcenygame/foxThief.t3dm");
+        players[playerNumber].model = t3d_model_load("rom:/rippergame/foxThief.t3dm");
     }
     else
     {
         players[playerNumber].playerTeam = teamGuard;
-        players[playerNumber].model = t3d_model_load("rom:/larcenygame/dogGuard.t3dm");
+        players[playerNumber].model = t3d_model_load("rom:/rippergame/dogGuard.t3dm");
     }
 
     switch(playerNumber)
     {
         case 0:
             players[playerNumber].playerPos = (T3DVec3){{-128, 0.0f, 128}};
-            players[playerNumber].rotY = 3.14f;
         break;
         case 1:
             players[playerNumber].playerPos = (T3DVec3){{128, 0.0f, 128}};
-            players[playerNumber].rotY = 3.14f;
         break;
         case 2:
             players[playerNumber].playerPos = (T3DVec3){{128, 0.0f, -128}};
-            players[playerNumber].rotY = 0.0f;
         break;
         case 3:
             players[playerNumber].playerPos = (T3DVec3){{-128, 0.0f, -128}};
-            players[playerNumber].rotY = 0.0f;
         default:
         break;
     }
@@ -864,6 +613,7 @@ void player_init(int playerNumber)
         t3d_matrix_pop(1);
     players[playerNumber].dplPlayer = rspq_block_end();
     players[playerNumber].moveDir = (T3DVec3){{0,0,0}};
+    players[playerNumber].rotY = 0.0f;
     players[playerNumber].currSpeed = 0.0f;
     players[playerNumber].animBlend = 0.0f;
     players[playerNumber].stunTimer = 0.0f;
@@ -898,7 +648,7 @@ void player_cleanup(int playerNumber)
 
 void player_fixedloop(float deltaTime, int playerNumber)
 {
-    if(gameStarting || gameEnding || gamePaused)
+    if(gameStarting || gameEnding)
     {
         return;
     }
@@ -960,14 +710,13 @@ void player_fixedloop(float deltaTime, int playerNumber)
 
     if(players[playerNumber].isAi && !(players[playerNumber].stunTimer > 0.0f)) // is an AI
     {
-        // TODO: debug stuff
-        //uint64_t aiStart = get_ticks();
         // run the generic ai update function to let the AI state machine worry about it
         newDir.x = 0.0f; newDir.y = 0.0f; newDir.z = 0.0f; speed = 0.0f;
         ai_update(players[playerNumber].aiIndex, deltaTime, &newDir, &speed);
-        // TODO: debug stuff
-        //aiTime += get_ticks() - aiStart;
     }
+
+    //TODO: Remove
+    //if(playerNumber == 0) tempSpeed = speed;
 
     if(speed > 0.15f)
     {
@@ -989,17 +738,20 @@ void player_fixedloop(float deltaTime, int playerNumber)
     tempPosition.v[0] += players[playerNumber].moveDir.v[0] * players[playerNumber].currSpeed;
     tempPosition.v[2] += players[playerNumber].moveDir.v[2] * players[playerNumber].currSpeed;
 
+    //TODO: remove this
+    //if(playerNumber == 0) tempIntersectionPoint = (T3DVec3){{0}};
+
     // do collision checks here
     collisionresult_data collisionResult;
-    
-    collision_check_intersect(&collisionResult, &players[playerNumber].playerPos, &tempPosition);
-    
-    // use collisionResult data to determine if to apply simulated move
-    if(collisionResult.didCollide && !(collisionResult.intersectionPoint.v[0] == 0.0f && collisionResult.intersectionPoint.v[2] == 0.0f))
-    {
-        players[playerNumber].playerPos = collisionResult.intersectionPoint;
-    }
 
+    collision_check(&collisionResult, &tempPosition);
+    
+    //collision_check_intersect(&collisionResult, &players[playerNumber].playerPos, &tempPosition);
+
+    //TODO: remove this
+    //if(playerNumber == 0) tempIntersectionPoint = collisionResult.intersectionPoint;
+
+    // use collisionResult data to determine if to apply simulated move
     if(!collisionResult.didCollide)
     {
         players[playerNumber].playerPos = tempPosition;
@@ -1038,12 +790,11 @@ void player_fixedloop(float deltaTime, int playerNumber)
             // get the vector offset between the player and the objective positions
             t3d_vec3_diff(&tempVec, &players[playerNumber].playerPos, &objectives[iDx].objectivePos);
             // if objective is closer than 10 units, then count as having been completed
-            if(t3d_vec3_len(&tempVec) < OBJECTIVE_TOUCH_DISTANCE)
+            if(t3d_vec3_len(&tempVec) < 10)
             {
                 objectives[iDx].isActive = false;
                 
                 wav64_play(&sfx_objectiveCompleted, 30);
-                mixer_ch_set_vol(30, 0.5f, 0.5f);
             }
         }
     }
@@ -1064,8 +815,7 @@ void player_fixedloop(float deltaTime, int playerNumber)
                 players[iDx].isActive = false;
 
                 
-                wav64_play(&sfx_thiefCaught, 27);
-                mixer_ch_set_vol(27, 0.5f, 0.5f);
+                wav64_play(&sfx_thiefCaught, 28);
             }
         }
     }
@@ -1078,7 +828,6 @@ void player_fixedloop(float deltaTime, int playerNumber)
 
 void player_guardAbility(float deltaTime, int playerNumber)
 {
-    bool tempHasHitSomeone = false;
     // stun in an AoE
     for(int iDx = 0; iDx < MAXPLAYERS; iDx++)
     {
@@ -1088,11 +837,7 @@ void player_guardAbility(float deltaTime, int playerNumber)
         }
         T3DVec3 tempVec = {0};
         t3d_vec3_diff(&tempVec, &players[playerNumber].playerPos, &players[iDx].playerPos);
-        if(t3d_vec3_len(&tempVec) < GUARD_ABILITY_RANGE && players[iDx].stunTimer == 0.0f) 
-        {
-            players[iDx].stunTimer = 1.0f;
-            tempHasHitSomeone = true;
-        }
+        if(t3d_vec3_len(&tempVec) < 50 && players[iDx].stunTimer == 0.0f) players[iDx].stunTimer = 1.0f;
     }
     // set the ability timer
     players[playerNumber].abilityTimer = DEFAULT_ABILITY_COOLDOWN;
@@ -1101,11 +846,8 @@ void player_guardAbility(float deltaTime, int playerNumber)
     t3d_anim_set_playing(&players[playerNumber].animAbility, true);
     t3d_anim_set_time(&players[playerNumber].animAbility, 0.0f);
 
-    // play sound effects here
+    // play sound effect here
     wav64_play(&sfx_guardStunAbility, 29);
-    mixer_ch_set_vol(29, 0.5f, 0.5f);
-    if(tempHasHitSomeone) wav64_play(&sfx_guardStunAbilityHit, 28);
-    mixer_ch_set_vol(28, 0.5f, 0.5f);
 }
 
 /*==============================
@@ -1123,7 +865,7 @@ void player_thiefAbility(float deltaTime, int playerNumber)
     tempUnitisedRotatedVector.v[0] = -sinf(players[playerNumber].rotY);
     tempUnitisedRotatedVector.v[2] = cosf(players[playerNumber].rotY);
 
-    for(float fDx = 1.0f; fDx < THIEF_ABILITY_RANGE; fDx+= THIEF_ABILITY_STRIDE)
+    for(float fDx = 1.0f; fDx < 24; fDx+= 2.0)
     {
         t3d_vec3_scale(&tempvec, &tempUnitisedRotatedVector, fDx);
         t3d_vec3_add(&tempvec, &tempvec, &players[playerNumber].playerPos);
@@ -1149,33 +891,61 @@ void player_thiefAbility(float deltaTime, int playerNumber)
             effectPool[effectPoolIndex].remainingTimer = 0.25f;
             effectPool[effectPoolIndex].effectPos = players[playerNumber].playerPos;
             effectPool[effectPoolIndex].effectRotationY = players[playerNumber].rotY;
-            effectPool[effectPoolIndex].effectSize = 30.0f;
+            effectPool[effectPoolIndex].effectSize = 20.0f;
 
             // make a second effect on the other side of the wall
             effectPoolIndex = effect_getNextEmptyIndex();
             effectPool[effectPoolIndex].isActive = true;
             effectPool[effectPoolIndex].remainingTimer = 0.25f;
             effectPool[effectPoolIndex].effectPos = tempvec;
-            effectPool[effectPoolIndex].effectRotationY = players[playerNumber].rotY + 3.14f; // flip the second
-            effectPool[effectPoolIndex].effectSize = 30.0f;
+            effectPool[effectPoolIndex].effectRotationY = players[playerNumber].rotY;
+            effectPool[effectPoolIndex].effectSize = 20.0f;
 
             // move the player
+            // TODO: make it a destination and do an animation over to it
             players[playerNumber].playerPos = tempvec;
 
             // play sound effect here
-            wav64_play(&sfx_thiefJumpAbility, 26);
-            mixer_ch_set_vol(26, 0.5f, 0.5f);
+            // TODO: make this its own sound effect
+            wav64_play(&sfx_guardStunAbility, 29);
 
             break;
         }
     }
+
+
+    // take out current position
+    // Get current location
+        //get a vector of 1,1
+        // X = Cos(Th), Y = Sin(Th)
+        // ex.  Cos(45) = 0.707, Sin(45) = 0.707
+        // rotation Th in degrees (not rads)
+        // result should be a unitised vector for direction
+        // can use this to iterate
+	// Get direction
+	// loop like X amount of times with difference between current location and direction times length
+	// see if the spot is open, if it is, set is as a destination and teleport to the other side
+    // if teleported, set abilitytimer
+
+    // scratchpad
+    // starting position: 45,10
+    // rotation of 45 degrees
+    // unitised rotated vector = 0.707, 0.707
+    // starting position += (URV * steps)
+    // 45.000, 10.000
+    // 45.707, 10.707
+    // 46.414, 11.414
+    // 47.121, 12.121
+    // 47.828, 12.828
+    // 48.535, 13.535
+    // 49.242, 14.242
+    // 49.949, 14.949
+    // 50.656, 15.656
+    // 51.363, 16.363
+    // 52.070, 17.070
 }
 
-/*==============================
-    player_stopAnimations
-    stops player animations, usually called when game ends
-==============================*/
-
+// stops player animations, usually called when game ends
 void player_stopAnimations(float deltaTime, int playerNumber)
 {
     players[playerNumber].animBlend = 0.0f;
@@ -1187,6 +957,8 @@ void player_stopAnimations(float deltaTime, int playerNumber)
     if(players[playerNumber].animAbility.isPlaying) t3d_anim_update(&players[playerNumber].animAbility, deltaTime);
 
     t3d_skeleton_blend(&players[playerNumber].skel, &players[playerNumber].skel, &players[playerNumber].skelBlend, players[playerNumber].animBlend);
+
+    if(syncPoint)rspq_syncpoint_wait(syncPoint); // wait for the RSP to process the previous frame
 
     t3d_skeleton_update(&players[playerNumber].skel);
     // update matrix
@@ -1207,7 +979,6 @@ void player_loop(float deltaTime, int playerNumber)
     if(!players[playerNumber].isActive) return;
 
     joypad_port_t controllerPort = core_get_playercontroller(playerNumber);
-    joypad_buttons_t btn;
     
     if(!joypad_is_connected(controllerPort))
     {
@@ -1219,56 +990,23 @@ void player_loop(float deltaTime, int playerNumber)
         player_stopAnimations(deltaTime, playerNumber);
     }
 
-    if(gamePaused)
-    {
-        // first pause all animations
-        t3d_anim_set_speed(&players[playerNumber].animIdle, 0.0f);
-        t3d_anim_set_speed(&players[playerNumber].animWalk, 0.0f);
-        t3d_anim_set_speed(&players[playerNumber].animAbility, 0.0f);
-
-        // check for someone pressing start to unpause
-        btn = joypad_get_buttons_held(controllerPort);
-        if(btn.start && gamePauseDebounce <= 0.0f)
-        {
-            gamePaused = false;
-            gamePauseDebounce = CONTROLLER_DEFAULT_DEBOUNCE;
-        }
-
-        if(btn.l && gamePauseDebounce <= 0.0f)
-        {
-            minigame_end();
-        }
-    }
-    else
-    {
-        t3d_anim_set_speed(&players[playerNumber].animIdle, 1.0f);
-        t3d_anim_set_speed(&players[playerNumber].animWalk, players[playerNumber].animBlend);
-        t3d_anim_set_speed(&players[playerNumber].animAbility, 1.0f);
-    }
-
     // TODO: Remove this because it's fucked
     // spam the fuck out of the button
     if(players[playerNumber].isAi && players[playerNumber].playerTeam == teamThief)
     {
-        if(!(players[playerNumber].abilityTimer > 0.0f) && !(players[playerNumber].stunTimer > 0.0f))
+        if(!(players[playerNumber].abilityTimer > 0.0f))
         {
             player_thiefAbility(deltaTime, playerNumber);
         }
     }
 
-    t3d_anim_update(&players[playerNumber].animIdle, deltaTime);
-    t3d_anim_update(&players[playerNumber].animWalk, deltaTime);
-    
-
-    if(!players[playerNumber].isAi && !gameStarting && !gameEnding && !gamePaused)
+    if(!players[playerNumber].isAi && !gameStarting && !gameEnding)
     {
-        btn = joypad_get_buttons_held(controllerPort);
+        joypad_buttons_t btn = joypad_get_buttons_held(controllerPort);
 
-        if(btn.start && gamePauseDebounce <= 0.0f)
-        {
-            gamePauseDebounce = CONTROLLER_DEFAULT_DEBOUNCE;
-            gamePaused = true;
-        }
+        if(btn.start) minigame_end();
+        if(btn.c_left) end_game(teamThief);
+        if(btn.c_right) end_game(teamGuard);
 
         // if A button pressed and player team is guard, then use guard ability
         if((btn.a || btn.b) && players[playerNumber].playerTeam == teamGuard)
@@ -1284,34 +1022,31 @@ void player_loop(float deltaTime, int playerNumber)
         if((btn.a || btn.b) && players[playerNumber].playerTeam == teamThief)
         {
             // check to see if our ability cooldown is not active
-            if(!(players[playerNumber].abilityTimer > 0.0f) && !(players[playerNumber].stunTimer > 0.0f))
+            if(!(players[playerNumber].abilityTimer > 0.0f))
             {
                 player_thiefAbility(deltaTime, playerNumber);
             }
         }
     }
 
-    if(players[playerNumber].animAbility.isPlaying)
-    {
-        t3d_anim_update(&players[playerNumber].animAbility, deltaTime);
-        players[playerNumber].animBlend = 1.0f;
-    }
+    t3d_anim_update(&players[playerNumber].animIdle, deltaTime);
+    t3d_anim_set_speed(&players[playerNumber].animIdle, 1.0f);
+    t3d_anim_set_speed(&players[playerNumber].animWalk, players[playerNumber].animBlend);
+    t3d_anim_update(&players[playerNumber].animWalk, deltaTime);
+
+    if(players[playerNumber].animAbility.isPlaying) t3d_anim_update(&players[playerNumber].animAbility, deltaTime);
 
     t3d_skeleton_blend(&players[playerNumber].skel, &players[playerNumber].skel, &players[playerNumber].skelBlend, players[playerNumber].animBlend);
 
-    t3d_skeleton_update(&players[playerNumber].skel);
+    if(syncPoint)rspq_syncpoint_wait(syncPoint); // wait for the RSP to process the previous frame
 
+    t3d_skeleton_update(&players[playerNumber].skel);
     // update matrix
     t3d_mat4fp_from_srt_euler(players[playerNumber].modelMatFP,
         (float[3]){DEFAULT_PLAYER_SCALE, DEFAULT_PLAYER_SCALE, DEFAULT_PLAYER_SCALE},
         (float[3]){0.0f, players[playerNumber].rotY, 0},
         players[playerNumber].playerPos.v);
 }
-
-/*==============================
-    player_draw
-    simple function to draw the chosen player
-==============================*/
 
 void player_draw(int playerNumber)
 {
@@ -1320,75 +1055,45 @@ void player_draw(int playerNumber)
     rspq_block_run(players[playerNumber].dplPlayer);
 }
 
-/*==============================
-    objective_init
-    Initialises and configures the chosen objectives
-==============================*/
-
 void objective_init()
 {
     objectives[0].isActive = true;
     objectives[0].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    objectives[0].objectiveModel = t3d_model_load("rom:/larcenygame/diamondObjective.t3dm");
-    objectives[0].ringModel = t3d_model_load("rom:/larcenygame/ringObjective.t3dm");
+    objectives[0].model = t3d_model_load("rom:/rippergame/testObjective.t3dm");
     rspq_block_begin();
         t3d_matrix_push(objectives[0].modelMatFP);
-        rdpq_set_prim_color(RGBA32(255,255,255,255));
-        t3d_model_draw(objectives[0].objectiveModel);
-        rdpq_set_prim_color(PLAYERCOLOR_4);
-        t3d_model_draw(objectives[0].ringModel);
+        rdpq_set_prim_color(RGBA32(0,255,0,255));
+        t3d_model_draw(objectives[0].model);
         t3d_matrix_pop(1);
     objectives[0].dplObjective = rspq_block_end();
-    objectives[0].objectiveRotationY = 0.0f;
-    objectives[0].objectivePos = (T3DVec3){{-70.0f, 0.0f, -70.0f}};
+    //objectives[0].objectivePos = (T3DVec3){{96, 0.0f, 96}};
+    objectives[0].objectivePos = (T3DVec3){{0.0f, 0.0f, 128}};
     
     objectives[1].isActive = true;
     objectives[1].modelMatFP = malloc_uncached(sizeof(T3DMat4FP));
-    objectives[1].objectiveModel = t3d_model_load("rom:/larcenygame/theKilogramObjective.t3dm");
-    objectives[1].ringModel = t3d_model_load("rom:/larcenygame/ringObjective.t3dm");
+    objectives[1].model = t3d_model_load("rom:/rippergame/testObjective.t3dm");
     rspq_block_begin();
         t3d_matrix_push(objectives[1].modelMatFP);
-        rdpq_set_prim_color(RGBA32(255,255,255,255));
-        t3d_model_draw(objectives[1].objectiveModel);
-        rdpq_set_prim_color(PLAYERCOLOR_2);
-        t3d_model_draw(objectives[1].ringModel);
+        rdpq_set_prim_color(RGBA32(255,0,0,255));
+        t3d_model_draw(objectives[1].model);
         t3d_matrix_pop(1);
     objectives[1].dplObjective = rspq_block_end();
-    objectives[1].objectiveRotationY = 0.0f;
-    objectives[1].objectivePos = (T3DVec3){{70.0f, 0.0f, 70}};
+    //objectives[1].objectivePos = (T3DVec3){{-96, 0.0f, -96}};
+    objectives[1].objectivePos = (T3DVec3){{0.0f, 0.0f, -128}};
 }
 
-/*==============================
-    objective_update
-    Non-time critical update function for the objectives,
-    just rotates and transforms them
-==============================*/
-
-void objective_update(float deltaTime)
+void objective_update()
 {
-    if(objectives[0].isActive)
-    {
-    objectives[0].objectiveRotationY += deltaTime;
     // update matricies
     t3d_mat4fp_from_srt_euler(objectives[0].modelMatFP,
         (float[3]){0.125f, 0.125f, 0.125f},
-        (float[3]){0.0f, objectives[0].objectiveRotationY, 0.0f},
+        (float[3]){0.0f, 0.0f, 0.0f},
         objectives[0].objectivePos.v);
-    }
-    if(objectives[1].isActive)
-    {
-        objectives[1].objectiveRotationY += deltaTime;
-        t3d_mat4fp_from_srt_euler(objectives[1].modelMatFP,
-            (float[3]){0.125f, 0.125f, 0.125f},
-            (float[3]){0.0f, objectives[1].objectiveRotationY, 0.0f},
-            objectives[1].objectivePos.v);  
-    }
+    t3d_mat4fp_from_srt_euler(objectives[1].modelMatFP,
+        (float[3]){0.125f, 0.125f, 0.125f},
+        (float[3]){0.0f, 0.0f, 0.0f},
+        objectives[1].objectivePos.v);  
 }
-
-/*==============================
-    objective_draw
-    Simple drawing function of all objectives
-==============================*/
 
 void objective_draw()
 {
@@ -1397,28 +1102,16 @@ void objective_draw()
     if(objectives[1].isActive) rspq_block_run(objectives[1].dplObjective);
 }
 
-/*==============================
-    objective_cleanup
-    Cleans and frees all allocated memory
-==============================*/
-
 void objective_cleanup()
 {
     rspq_block_free(objectives[0].dplObjective);
-    t3d_model_free(objectives[0].objectiveModel);
-    t3d_model_free(objectives[0].ringModel);
+    t3d_model_free(objectives[0].model);
     free_uncached(objectives[0].modelMatFP);
 
     rspq_block_free(objectives[1].dplObjective);
-    t3d_model_free(objectives[1].objectiveModel);
-    t3d_model_free(objectives[1].ringModel);
+    t3d_model_free(objectives[1].model);
     free_uncached(objectives[1].modelMatFP);
 }
-
-/*==============================
-    effect_init
-    Initialises the pool of effects
-==============================*/
 
 void effect_init()
 {
@@ -1434,12 +1127,7 @@ void effect_init()
     }
 }
 
-/*==============================
-    effect_getNextEmptyIndex
-    returns the index of the first unused effect,
-    returns 0 (overwrites the first one) if none free
-==============================*/
-
+// returns the index of the first unused effect, returns 0 (overwrites the first one) if none free
 int effect_getNextEmptyIndex()
 {
     int tempReturnValue = 0;
@@ -1456,11 +1144,6 @@ int effect_getNextEmptyIndex()
     return tempReturnValue;
 }
 
-/*==============================
-    effect_update
-    Updates the full effects pool
-==============================*/
-
 void effect_update(float deltaTime)
 {
     for(int iDx = 0; iDx < MAXPLAYERS; iDx++)
@@ -1476,11 +1159,6 @@ void effect_update(float deltaTime)
         }
     }
 }
-
-/*==============================
-    effect_draw
-    Draws the full effects pool
-==============================*/
 
 void effect_draw()
 {
@@ -1500,11 +1178,6 @@ void effect_draw()
         }
     }
 }
-
-/*==============================
-    effect_cleanup
-    Free's and cleans all memory allocated by the effects
-==============================*/
 
 void effect_cleanup()
 {
@@ -1537,23 +1210,8 @@ void minigame_init()
     gameTimeRemaining = STARTING_GAME_TIME;
 
     // initialise the display, setting resolution, colour depth and AA
-    joypad_buttons_t btn;
-    btn = joypad_get_buttons_held(0);
-    if(btn.r)
-    {
-        resolutionDisplayX = 640;
-        resolutionDisplayY = 480;
-        display_init(RESOLUTION_640x480, DEPTH_16_BPP, 3, GAMMA_NONE, FILTERS_RESAMPLE);
-    }
-    else
-    {  
-        resolutionDisplayX = 320;
-        resolutionDisplayY = 240;
-        display_init(RESOLUTION_320x240, DEPTH_16_BPP, 3, GAMMA_NONE, FILTERS_RESAMPLE);
-    }
+    display_init(RESOLUTION_320x240, DEPTH_16_BPP, 3, GAMMA_NONE, FILTERS_RESAMPLE);
     depthBuffer = display_get_zbuf();
-
-    //rdpq_debug_start();
 
     // start tiny3d
     t3d_init((T3DInitParams){});
@@ -1561,8 +1219,6 @@ void minigame_init()
     // load a font to use for HUD text
     fontDebug = rdpq_font_load_builtin(FONT_BUILTIN_DEBUG_MONO);
     rdpq_text_register_font(FONT_DEBUG, fontDebug);
-
-    rdpq_font_style(fontDebug, 7, &(rdpq_fontstyle_t){ .color = RGBA32(255,255,255,255) });
 
     // load in the player billboard font
     fontBillboard = rdpq_font_load("rom:/squarewave.font64");
@@ -1579,19 +1235,16 @@ void minigame_init()
     mapMatFP = malloc_uncached(sizeof(T3DMat4FP));
     t3d_mat4fp_from_srt_euler(mapMatFP, (float[3]){0.3f, 0.3f, 0.3f}, (float[3]){0, 0, 0}, (float[3]){0,0,-10});
 
-    // set camera position and target vectors to defaults
-    camPos = (T3DVec3){{0, 255.0f, 45.0f}};
-    camTarget = (T3DVec3){{0, 0, -5}};
-
-    // init the camera system
-    cameraAnimation_init();
+    // set camera position and target vectors
+    camPos = (T3DVec3){{0, 215.0f, 65.0f}};
+    camTarget = (T3DVec3){{0, 0, 20}};
 
     // set up a vector for the directional light
     lightDirVec = (T3DVec3){{1.0f, 1.0f, 1.0f}};
     t3d_vec3_norm(&lightDirVec);
 
     // load a model from ROM for the map
-    modelMap = t3d_model_load("rom:/larcenygame/map.t3dm");
+    modelMap = t3d_model_load("rom:/rippergame/map.t3dm");
 
     // create a command block/display list to optimise drawing the map
     rspq_block_begin();
@@ -1602,10 +1255,10 @@ void minigame_init()
     dplMap = rspq_block_end();
 
     // load a model from ROM for the stun weapon effect
-    modelWallJumpEffect = t3d_model_load("rom:/larcenygame/dustEffect.t3dm");
+    modelWallJumpEffect = t3d_model_load("rom:/rippergame/dustEffect.t3dm");
 
     // load a model from ROM for the collision square
-    modelCollision = t3d_model_load("rom:/larcenygame/collisionSquare.t3dm");
+    modelCollision = t3d_model_load("rom:/rippergame/collisionSquare.t3dm");
 
     // load sprites from ROM for the HUD
     spriteAButton = sprite_load("rom:/core/AButton.sprite");
@@ -1618,14 +1271,14 @@ void minigame_init()
     wav64_open(&sfx_countdown, "rom:/core/Countdown.wav64");
     wav64_open(&sfx_winner, "rom:/core/Winner.wav64");
     
-    // TODO: Own sounds
-    wav64_open(&sfx_objectiveCompleted, "rom:/larcenygame/objectiveTouch.wav64");
-    wav64_open(&sfx_guardStunAbility, "rom:/larcenygame/guardSwing.wav64");
-    wav64_open(&sfx_guardStunAbilityHit, "rom:/larcenygame/guardStunHit.wav64");
-    wav64_open(&sfx_thiefJumpAbility, "rom:/larcenygame/thiefJump.wav64");
-    wav64_open(&sfx_thiefCaught, "rom:/larcenygame/thiefCaught.wav64");
+    wav64_open(&sfx_objectiveCompleted, "rom:/core/Countdown.wav64");
+    wav64_open(&sfx_guardStunAbility, "rom:/core/Countdown.wav64");
+    wav64_open(&sfx_thiefJumpAbility, "rom:/core/Countdown.wav64");
+    wav64_open(&sfx_thiefCaught, "rom:/core/Countdown.wav64");
 
-    xm64player_open(&xm_music, "rom:/larcenygame/drozerix-poppy_flower_girls.xm64");
+    xm64player_open(&xm_music, "rom:/snake3d/bottled_bubbles.xm64");
+    
+    mixer_ch_set_vol(31, 0.5f, 0.5f);
 
     // ensure AI init is done before player_init, otherwise no AI to be assigned to
     ai_init(players, MAXPLAYERS, objectives, sizeof(objectives)/sizeof(objectives[0]), 
@@ -1646,9 +1299,6 @@ void minigame_init()
     // initialise effect models
     effect_init();
 
-    // set off the intro animation system
-    startCameraAnimationIntro();
-
     return;
 }
 
@@ -1660,67 +1310,48 @@ void minigame_init()
     @param  The fixed delta time for this tick
 ==============================*/
 
-void minigame_fixedloop(float deltaTime)
+void minigame_fixedloop(float deltatime)
 {
-    // TODO: fixedloop update debug stuff
-    //fixedUpdateTime = 0;
-    //aiTime = 0;
-    //uint64_t fixedUpdateStart = get_ticks();
-    
     // update the player entities
     for(int i = 0; i < MAXPLAYERS; i++)
     {
-        player_fixedloop(deltaTime, i);
+        player_fixedloop(deltatime, i);
     }
 
-    HUD_Update(deltaTime);
+    HUD_Update(deltatime);
 
-    effect_update(deltaTime);
+    effect_update(deltatime);
 
     // process pre-game countdown timer
     if(gameStarting)
     {
-        countdownTimer -= deltaTime / 2.5;
+        countdownTimer -= deltatime;
 
         if(countdownTimer < lastCountdownNumber)
         {
             lastCountdownNumber = countdownTimer;
             wav64_play(&sfx_countdown, 31);
-            mixer_ch_set_vol(31, 0.5f, 0.5f);
         }
 
         if(countdownTimer < 1.0f)
         {
             wav64_play(&sfx_start, 31);
-            mixer_ch_set_vol(31, 0.75f, 0.75f);
             gameStarting = false;
 
             // TODO: start playing music
             xm64player_play(&xm_music, 0);
-            xm64player_set_loop(&xm_music, true);
-            xm64player_set_vol(&xm_music, 0.5f);
         }
-    }
-
-    // if there's any value in the debounce, make sure to address it
-    if(gamePauseDebounce >= 0.0f)
-    {
-        gamePauseDebounce -= deltaTime;
-        if(gamePauseDebounce < 0.0f) gamePauseDebounce = 0.0f;
     }
 
     if(gameEnding)
     {
-        countdownTimer -= deltaTime;
+        countdownTimer -= deltatime;
 
         if(countdownTimer < 1.0f)
         {
             minigame_end();
         }
     }
-    
-    // TODO: debug stuff
-    //fixedUpdateTime += get_ticks() - fixedUpdateStart;
     return;
 }
 
@@ -1730,37 +1361,20 @@ void minigame_fixedloop(float deltaTime)
     @param  The delta time for this tick
 ==============================*/
 
-void minigame_loop(float deltaTime)
+void minigame_loop(float deltatime)
 {
-    // TODO: Debug timers
-    // reset timers
-    /*startTime = 0, endTime = 0;
-    updateTime = 0;
-    colTime = 0;
-    playerUpdate = 0;
-    drawTime = 0;
-    subDrawTime = 0;
-    totalFrameTime = 0;*/
-    // Start timing
-    //startTime = get_ticks();
-    
-    if(syncPoint)rspq_syncpoint_wait(syncPoint); // wait for the RSP to process the previous frame
 
-    // TODO: debug stuff
-    //uint64_t updateStart = get_ticks();
-    //uint64_t playerUpdateStart = get_ticks();
     // update the player entities
-    for(int iDx = 0; iDx < MAXPLAYERS; iDx++)
+    for(int i = 0; i < MAXPLAYERS; i++)
     {
-        player_loop(deltaTime, iDx);
+        player_loop(deltatime, i);
     }
 
     // run objective updates
-    objective_update(deltaTime);
-    //playerUpdate += get_ticks() - playerUpdateStart;
+    objective_update();
 
     // Check for victory conditions
-    if(!gameStarting && !gameEnding && !gamePaused)
+    if(!gameStarting && !gameEnding)
     {
         if(gameTimeRemaining < 0) end_game(teamGuard);
         // start this bool as true and set it false as soon as you get a single active objective
@@ -1789,56 +1403,42 @@ void minigame_loop(float deltaTime)
         if(tempGameEndFlag) end_game(teamGuard);
     }
 
+    // set up the ambient light colour and the directional light colour
+    uint8_t colorAmbient[4] = {0xAA, 0xAA, 0xAA, 0xFF};
+    uint8_t colorDir[4] = {0xFF, 0xAA, 0xAA, 0xFF};
 
     // set the projection matrix for the viewport
-    //float aspectRatio = resolutionDisplayX / resolutionDisplayY;
     float aspectRatio = (float)viewport.size[0] / ((float)viewport.size[1]);//*2);
-    t3d_viewport_set_perspective(&viewport, T3D_DEG_TO_RAD(90.0f), aspectRatio, 20.0f, 600.0f);
-    // update the camera if needed
-    cameraAnimation_update(deltaTime);
+    t3d_viewport_set_perspective(&viewport, T3D_DEG_TO_RAD(90.0f), aspectRatio, 20.0f, 260.0f);
     t3d_viewport_look_at(&viewport, &camPos, &camTarget, &(T3DVec3){{0,1,0}});
 
-    // TODO: debug stuff
-    //updateTime += get_ticks() - updateStart;
-    // Before drawing anything start the timer
-    //uint64_t drawStart = get_ticks();
     // Draw our 3D frame
     // grab the colour and depth buffers and attach
     // the rdp queue in order to dispatch commands to the RDP
     rdpq_attach(display_get(), depthBuffer);
-
-    // set up the ambient light colour and the directional light colour
-    uint8_t colorAmbient[4] = {0xAA, 0xAA, 0xAA, 0xFF};
-    uint8_t colorDir[4] = {0xFF, 0xAA, 0xAA, 0xFF};
-    
-    // set up a vector for the directional light
-    lightDirVec = (T3DVec3){{1.0f, 1.0f, 1.0f}};
-    t3d_vec3_norm(&lightDirVec);
 
     // tell tiny3d it's time to start a new frame
     t3d_frame_start();
     // attach the desired viewport
     t3d_viewport_attach(&viewport);
 
+    // clear the colour and depth buffers
+    t3d_screen_clear_color(RGBA32(224, 180, 96, 0xFF));
+    t3d_screen_clear_depth();
+
     // set the lighting details
     t3d_light_set_ambient(colorAmbient);
     t3d_light_set_directional(0, colorDir, &lightDirVec);
     t3d_light_set_count(1);
 
-    // clear the colour and depth buffers
-    //t3d_screen_clear_color(RGBA32(224, 180, 96, 0xFF));
-    t3d_screen_clear_color(RGBA32(32, 32, 32, 0xFF));
-    t3d_screen_clear_depth();
-
     // draw collision squares
-    //collision_draw();
+    collision_draw();
 
     // draw the player entities
     for(int i = 0; i < MAXPLAYERS; i++)
     {
         player_draw(i);
     }
-
 
     // run the displaylist containing the map draw routine
     rspq_block_run(dplMap);
@@ -1848,36 +1448,37 @@ void minigame_loop(float deltaTime)
 
     // draw the effects
     effect_draw();
-    
+
+    // set a sync point
+    syncPoint = rspq_syncpoint_new();
+
+    // draws RDP text on top of the scene showing debug info
+    debugInfoDraw(deltatime);
+
     // draws the main game HUD
-    if(!gameStarting && !gameEnding && !gamePaused) HUD_draw();
+    HUD_draw();
+
+    // make sure the RDP is sync'd
+    rdpq_sync_tile();
+    rdpq_sync_pipe(); // Hardware crashes otherwise
 
     // game starting countdown text draw
     if(gameStarting)
     {
-        rdpq_textparms_t textparms = { .style_id = 7, .align = ALIGN_CENTER, .width = resolutionDisplayX, .disable_aa_fix = true, };
-        rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
-        rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, resolutionDisplayY / 2, "Starting in %i...", (int)countdownTimer);
-    }
-
-    if(gamePaused)
-    {
-        rdpq_textparms_t textparms = { .style_id = 7, .align = ALIGN_CENTER, .width = resolutionDisplayX, .disable_aa_fix = true, };
-        rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
-        rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, resolutionDisplayY / 2, "Game Paused \n Press L to Quit");
+        rdpq_textparms_t textparms = { .align = ALIGN_CENTER, .width = RESOLUTION_WIDTH, .disable_aa_fix = true, };
+        rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, 117, "Starting in %i...", (int)countdownTimer);
     }
 
     if(gameEnding)
     {
-        rdpq_textparms_t textparms = { .style_id = 7, .align = ALIGN_CENTER, .width = resolutionDisplayX, .disable_aa_fix = true, };
-        rdpq_sync_tile(); rdpq_sync_pipe(); // make sure the RDP is sync'd Hardware crashes otherwise
+        rdpq_textparms_t textparms = { .align = ALIGN_CENTER, .width = RESOLUTION_WIDTH, .disable_aa_fix = true, };
         if(winningTeam == teamThief)
         {
-            rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, resolutionDisplayY / 2, "Thieves Win!");
+            rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, 117, "Thieves Win!");
         }
         else
         {
-            rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, resolutionDisplayY / 2, "Guards Win!");
+            rdpq_text_printf(&textparms, FONT_BUILTIN_DEBUG_MONO, 0, 117, "Guards Win!");
         }
     }
     
@@ -1885,20 +1486,8 @@ void minigame_loop(float deltaTime)
     rdpq_sync_tile();
     rdpq_sync_pipe(); // Hardware crashes otherwise
     
-    //drawTime += get_ticks() - drawStart;
-
-    // End timing
-    endTime = get_ticks();
-    totalFrameTime = endTime - startTime;
-    
-    // draws RDP text on top of the scene showing debug info
-    //debugInfoDraw(deltaTime);
-    
     // detach the queue to flip the buffers and show it on screen
     rdpq_detach_show();
-
-    // set a sync point
-    syncPoint = rspq_syncpoint_new();
 
     return;
 }
@@ -1931,7 +1520,6 @@ void minigame_cleanup()
     wav64_close(&sfx_winner);
     wav64_close(&sfx_objectiveCompleted);
     wav64_close(&sfx_guardStunAbility);
-    wav64_close(&sfx_guardStunAbilityHit);
     wav64_close(&sfx_thiefJumpAbility);
     wav64_close(&sfx_thiefCaught);
