+++ code/larcenygame/larcenygame.mk
@@ -1,20 +1,17 @@
 
 ASSETS_LIST += \
-	filesystem/larcenygame/map.t3dm \
-	filesystem/larcenygame/collisionSquare.t3dm \
-	filesystem/larcenygame/foxThief.t3dm \
-	filesystem/larcenygame/dogGuard.t3dm \
-	filesystem/larcenygame/dustEffect.t3dm \
-	filesystem/larcenygame/ringObjective.t3dm \
-	filesystem/larcenygame/diamondObjective.t3dm \
-	filesystem/larcenygame/theKilogramObjective.t3dm \
-	filesystem/larcenygame/guardStunHit.wav64 \
-	filesystem/larcenygame/guardSwing.wav64 \
-	filesystem/larcenygame/thiefCaught.wav64 \
-	filesystem/larcenygame/thiefJump.wav64 \
-	filesystem/larcenygame/objectiveTouch.wav64 \
-	filesystem/larcenygame/objectiveGradient.ia8.sprite \
-	filesystem/larcenygame/thiefEye.ci4.sprite \
-	filesystem/larcenygame/floorTest.ci4.sprite \
-	filesystem/larcenygame/wallTest.ci4.sprite \
-	filesystem/larcenygame/drozerix-poppy_flower_girls.xm64
+	filesystem/rippergame/map.t3dm \
+	filesystem/rippergame/collisionSquare.t3dm \
+	filesystem/rippergame/foxThief.t3dm \
+	filesystem/rippergame/dogGuard.t3dm \
+	filesystem/rippergame/dustEffect.t3dm \
+	filesystem/rippergame/testActor.t3dm \
+	filesystem/rippergame/testActorGuard.t3dm \
+	filesystem/rippergame/testObjective.t3dm \
+	filesystem/rippergame/objectiveGradient.ia8.sprite \
+	filesystem/rippergame/sand12.ci4.sprite \
+	filesystem/rippergame/stone.ci4.sprite \
+	filesystem/rippergame/test32.ci4.sprite \
+	filesystem/rippergame/test32Inverted.ci4.sprite \
+	filesystem/rippergame/floorTest.ci4.sprite \
+	filesystem/rippergame/wallTest.ci4.sprite
