+++ code/larcenygame/larcenygameAI.c
@@ -1,4 +1,4 @@
-#include "./larcenygameAI.h"
+#include "./rippergameAI.h"
 
 
 /*********************************
@@ -33,22 +33,11 @@ int stunAbilityUseDistance;
         State functions
 *********************************/
 
-/*==============================
-    ai_waitingStateEnter
-    Called to enter the waiting state,
-    picks a random amount of time to wait
-==============================*/
-
 void ai_waitingStateEnter(int aiIndex)
 {
     aiData[aiIndex].currentState = StateWaiting;
     aiData[aiIndex].framesRemainingBeforeCheck = rand()%30;
 }
-
-/*==============================
-    ai_waitingState
-    Main logic for the waiting state
-==============================*/
 
 void ai_waitingState(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed)
 {
@@ -59,22 +48,12 @@ void ai_waitingState(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed
     }
     else
     {
-        if(!ai_checkForProximityBasedStateChanges(aiIndex))
-        {
-            *newDir = playersRef[aiIndex].playerPos;
-            *speed = 0.0f;
+        *newDir = playersRef[aiIndex].playerPos;
+        *speed = 0.0f;
 
-            aiData[aiIndex].framesRemainingBeforeCheck--;
-        }
+        aiData[aiIndex].framesRemainingBeforeCheck--;
     }
 }
-
-/*==============================
-    ai_wanderingStateEnter
-    Called to enter the wandering state,
-    picks a random amount of time to wait
-    as well as a random direction and speed
-==============================*/
 
 void ai_wanderingStateEnter(int aiIndex)
 {
@@ -84,11 +63,6 @@ void ai_wanderingStateEnter(int aiIndex)
     aiData[aiIndex].destination.v[0] = (rand()%170-85) * 0.05f;
     aiData[aiIndex].destination.v[2] = (rand()%170-85) * 0.05f;
 }
-
-/*==============================
-    ai_wanderingState
-    Main logic for the wandering state
-==============================*/
 
 void ai_wanderingState(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed)
 {
@@ -113,12 +87,6 @@ void ai_wanderingState(int aiIndex, float deltaTime, T3DVec3* newDir, float* spe
     }
 }
 
-/*==============================
-    ai_followingOtherPlayerStateEnter
-    Called to enter the following state,
-    finds the closest other player to follow
-==============================*/
-
 void ai_followingOtherPlayerStateEnter(int aiIndex)
 {
     aiData[aiIndex].currentState = StateFollowingOtherPlayer;
@@ -134,11 +102,6 @@ void ai_followingOtherPlayerStateEnter(int aiIndex)
     aiData[aiIndex].targetIndex = returnedStruct.targetIndex;
     aiData[aiIndex].targetType = returnedStruct.targetType;
 }
-
-/*==============================
-    ai_followingOtherPlayerState
-    Main logic for the following state
-==============================*/
 
 void ai_followingOtherPlayerState(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed)
 {
@@ -202,12 +165,6 @@ void ai_followingOtherPlayerState(int aiIndex, float deltaTime, T3DVec3* newDir,
     if(*speed > speedModifierDifficulty) *speed = speedModifierDifficulty;
 }
 
-/*==============================
-    ai_moveToObjectiveStateEnter
-    Called to enter the moveToObjective state,
-    finds the closest objective to move to and camp
-==============================*/
-
 void ai_moveToObjectiveStateEnter(int aiIndex)
 {
     //TODO: Finish this function
@@ -219,11 +176,6 @@ void ai_moveToObjectiveStateEnter(int aiIndex)
     aiData[aiIndex].targetIndex = returnedStruct.targetIndex;
     aiData[aiIndex].targetType = returnedStruct.targetType;
 }
-
-/*==============================
-    ai_moveToObjectiveState
-    Main logic for the moveToObjective state
-==============================*/
 
 void ai_moveToObjectiveState(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed)
 {
@@ -237,10 +189,7 @@ void ai_moveToObjectiveState(int aiIndex, float deltaTime, T3DVec3* newDir, floa
     // if the timer has elapsed, find us a new thing to do
     if(aiData[aiIndex].framesRemainingBeforeCheck <= 0)
     {
-        if(!ai_checkForProximityBasedStateChanges(aiIndex))
-        {
-            ai_findNewState(aiIndex);
-        }
+        ai_findNewState(aiIndex);
         return;
     }
     else
@@ -272,12 +221,6 @@ void ai_moveToObjectiveState(int aiIndex, float deltaTime, T3DVec3* newDir, floa
     }
 }
 
-/*==============================
-    ai_runningFromGuardStateEnter
-    Called to enter the runningFromGuard state,
-    finds the closest guard and begins to run away
-==============================*/
-
 void ai_runningFromGuardStateEnter(int aiIndex)
 {
     aiData[aiIndex].currentState = StateRunningFromGuard;
@@ -297,11 +240,6 @@ void ai_runningFromGuardStateEnter(int aiIndex)
     aiData[aiIndex].targetIndex = returnedStruct.targetIndex;
     aiData[aiIndex].targetType = returnedStruct.targetType;
 }
-
-/*==============================
-    ai_runningFromGuardState
-    Main logic for the runningFromGuard state
-==============================*/
 
 void ai_runningFromGuardState(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed)
 {
@@ -362,12 +300,6 @@ void ai_runningFromGuardState(int aiIndex, float deltaTime, T3DVec3* newDir, flo
     if(*speed > speedModifierDifficulty) *speed = speedModifierDifficulty;
 }
 
-/*==============================
-    ai_findClosestEntityOfType
-    Is passed in a type of entity to look for
-    and returns its information in the returnStruct
-==============================*/
-
 void ai_findClosestEntityOfType(EntitySearchReturnData* returnStruct, int aiIndex, AITargetType desiredType)
 {
     // first zero out the returnStruct
@@ -422,12 +354,7 @@ void ai_findClosestEntityOfType(EntitySearchReturnData* returnStruct, int aiInde
     return;
 }
 
-/*==============================
-    ai_checkForProximityBasedStateChanges
-    Checks for if guard/thieves are too close to each other,
-    returns true if state changed, false if not
-==============================*/
-
+// returns true if state changed, false if not
 bool ai_checkForProximityBasedStateChanges(int aiIndex)
 {
     EntitySearchReturnData returnedStruct;
@@ -464,12 +391,7 @@ bool ai_checkForProximityBasedStateChanges(int aiIndex)
     return false;
 }
 
-/*==============================
-    ai_checkForProximityBasedStateChanges
-    Handles randomly selecting a new state to enter,
-    puts all logic for it in one place
-==============================*/
-
+// handles randomly selecting a new state to enter
 void ai_findNewState(int aiIndex)
 {
     // time for a new state, time to pick one randomly
@@ -486,12 +408,6 @@ void ai_findNewState(int aiIndex)
             break;
     }
 }
-
-/*==============================
-    ai_getCurrentStateAsInt
-    Returns an int that corresponds with 
-    the state, used for debug info only
-==============================*/
 
 int  ai_getCurrentStateAsInt(int aiIndex)
 {
@@ -524,12 +440,7 @@ int  ai_getCurrentStateAsInt(int aiIndex)
     return -1;
 }
 
-/*==============================
-    ai_init
-    Sets the local references to needed object arrays
-    pass in player, objective and collision pointers
-==============================*/
-
+// pass in player, objective and collision pointers
 void ai_init(player_data* a_players, int a_playerDataSize, objective_data* a_objectives, 
         int a_objectiveDataSize, collisionobject_data* a_collisionObjects, int a_collisionObjectSize)
 {
@@ -558,6 +469,7 @@ void ai_init(player_data* a_players, int a_playerDataSize, objective_data* a_obj
             break;
     }
     
+
     // initialise the AI data structs
     for(int iDx = 0; iDx < MAXPLAYERS; iDx++)
     {
@@ -574,23 +486,12 @@ void ai_init(player_data* a_players, int a_playerDataSize, objective_data* a_obj
     return;
 }
 
-/*==============================
-    ai_assign
-    Returns an index to the AI to be passed back in for ai_update
-==============================*/
-
-int ai_assign(int a_playerNumber)
+int ai_assign(int a_playerNumber) // returns an index to the AI to be passed back in for ai_update
 {
     aiData[a_playerNumber].controlledEntityIndex = a_playerNumber;
     aiData[a_playerNumber].isAnAI = true;
     return a_playerNumber;
 }
-
-/*==============================
-    ai_update
-    Checks the current state of the FSM and executes
-    the applicable functions
-==============================*/
 
 void ai_update(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed) // main update function that traverses the state machine, pass in index and deltaTime
 {
@@ -622,12 +523,7 @@ void ai_update(int aiIndex, float deltaTime, T3DVec3* newDir, float* speed) // m
     return;
 }
 
-/*==============================
-    ai_cleanup
-    Ensures anything that needs to be freed, is
-==============================*/
-
-void ai_cleanup() 
+void ai_cleanup() // ensures anything that needs to be freed, is
 {
     return;
 }
