--- code/rippergame/rippergame.mk
@@ -1,17 +1,0 @@
-
-ASSETS_LIST += \
-	filesystem/rippergame/map.t3dm \
-	filesystem/rippergame/collisionSquare.t3dm \
-	filesystem/rippergame/foxThief.t3dm \
-	filesystem/rippergame/dogGuard.t3dm \
-	filesystem/rippergame/dustEffect.t3dm \
-	filesystem/rippergame/testActor.t3dm \
-	filesystem/rippergame/testActorGuard.t3dm \
-	filesystem/rippergame/testObjective.t3dm \
-	filesystem/rippergame/objectiveGradient.ia8.sprite \
-	filesystem/rippergame/sand12.ci4.sprite \
-	filesystem/rippergame/stone.ci4.sprite \
-	filesystem/rippergame/test32.ci4.sprite \
-	filesystem/rippergame/test32Inverted.ci4.sprite \
-	filesystem/rippergame/floorTest.ci4.sprite \
-	filesystem/rippergame/wallTest.ci4.sprite
